Respect reduced motion setting
Remove the debounced search behavior and instead wait for the user to
submit the query. We wrap the input in a `form` element to facilitate
this, we can just listen for the `submit` event on the form.

Additionally, uses the `prefers-reduced-motion` media query to remove
the animation on the search input when it is submitted. We leave the
other animations alone because they're far more subtle and reduced
motion doesn't necessarily mean _no_ motion.